requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pysyncobj==0.3.13
msgpack-python==0.5.6
prometheus-client==0.19.0
python-dotenv==1.0.0
httpx==0.26.0
pytest~=8.3.4
pytest-asyncio~=0.24.0
h11~=0.14.0
exceptiongroup~=1.2.2
sniffio~=1.3.1
click~=8.1.7
httpcore~=1.0.7
idna~=3.10
certifi~=2024.8.30
pluggy~=1.5.0
iniconfig~=2.0.0
pydantic~=2.10.2
anyio~=4.6.2.post1
annotated-types~=0.7.0
packaging~=24.2
tomli~=2.2.1